e2e_test_suite/requirements.txt
apache-beam[gcp]==2.57.0
pytest==7.4.4
google-cloud-pubsub==2.18.4
google-cloud-bigquery==3.11.4
google-cloud-storage==2.16.0
python-dotenv==1.0.0